Move the Cat object to chisel3 package Move the location of the file Fix up imports Add backward compatible deprecated Cat that points to new Cat Fix up usages
