+++ src/main/scala/chisel3/Cat.scala
@@ -1,12 +1,9 @@
-// SPDX-License-Identifier: Apache-2.0
+package chisel3
 
-package chisel3.util
-
-import scala.language.experimental.macros
-
-import chisel3._
 import chisel3.experimental.SourceInfo
 import chisel3.internal.sourceinfo.SourceInfoTransform
+
+import scala.language.experimental.macros
 
 /** Concatenates elements of the input, in order, together.
   *
