+++ src/test/scala/chiselTests/CatSpec.scala
@@ -1,11 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
-package chiselTests.util
+package chiselTests
 
 import chisel3._
 import chisel3.experimental.noPrefix
-import chisel3.util.Cat
-import chiselTests.ChiselFlatSpec
 import circt.stage.ChiselStage
 
 object CatSpec {
@@ -23,7 +21,7 @@ class CatSpec extends ChiselFlatSpec {
 
   import CatSpec._
 
-  behavior.of("util.Cat")
+  behavior.of("chisel3.Cat")
 
   it should "not fail to elaborate a zero-element Vec" in {
